new plugin architechture
