+++ .travis.yml
@@ -1,15 +1,15 @@
+notifications:
+  email: false
 sudo: false
 language: go
 go:
-- '1.10'
-- tip
+- '1.12.x'
+
 install:
-- go get github.com/go-task/task/cmd/task
-script:
-- task dl-deps
-- task update-deps
-- task lint
-- task dist_all
+  - curl https://raw.githubusercontent.com/golang/dep/master/install.sh | sh
+
+before_deploy:
+  - make dist
 
 deploy:
   provider: releases
