--- Taskfile.yml
@@ -1,48 +1,0 @@
-default:
-  deps: [test]
-
-dl-deps:
-  cmds:
-    - go get -u github.com/golang/dep/cmd/dep
-
-update-deps:
-  cmds:
-    - dep ensure
-    - dep ensure -update
-
-lint:
-  cmds:
-    - go fmt ./...
-    - go vet ./...
-  silent: true
-
-dist_all:
-  cmds:
-    - task: _dist
-      vars: {OS: "linux", ARCH: "386", exeExt: ""}
-    - task: _dist
-      vars: {OS: "linux", ARCH: "amd64", exeExt: ""}
-    - task: _dist
-      vars: {OS: "darwin", ARCH: "386", exeExt: ""}
-    - task: _dist
-      vars: {OS: "darwin", ARCH: "amd64", exeExt: ""}
-    - task: _dist
-      vars: {OS: "windows", ARCH: "386", exeExt: ".exe"}
-    - task: _dist
-      vars: {OS: "windows", ARCH: "amd64", exeExt: ".exe"}
-
-dist:
-  cmds:
-    - task: _dist
-      vars: {OS: "{{OS}}", ARCH: "{{ARCH}}", exeExt: "{{exeExt}}"}
-
-_dist:
-  cmds:
-    - GOOS={{.OS}} GOARCH={{.ARCH}} go build -o plugin{{.exeExt}} ./cmd/emojigen/plugin
-    - mkdir -p dist/
-    - tar -czf dist/mattermost-emojigen_{{.OS}}_{{.ARCH}}.tar.gz plugin.json plugin{{.exeExt}}
-    - rm plugin{{.exeExt}}
-
-clean:
-  cmds:
-    - rm -r dist/
