+++ go.mod
@@ -1,0 +1,40 @@
+module github.com/kaakaa/mattermost-emojigen
+
+go 1.12
+
+require (
+	github.com/blang/semver v3.5.1+incompatible
+	github.com/davecgh/go-spew v1.1.1
+	github.com/dyatlov/go-opengraph v0.0.0-20180429202543-816b6608b3c8
+	github.com/fsnotify/fsnotify v1.4.7
+	github.com/go-ldap/ldap v3.0.3+incompatible
+	github.com/golang/protobuf v1.3.1
+	github.com/google/uuid v1.1.1
+	github.com/gorilla/websocket v1.4.0
+	github.com/hashicorp/go-hclog v0.9.2
+	github.com/hashicorp/go-plugin v1.0.0
+	github.com/hashicorp/yamux v0.0.0-20181012175058-2f1d1f20f75d
+	github.com/mattermost/go-i18n v1.10.0
+	github.com/mattermost/mattermost-server v0.0.0-20190519073129-18612bf771d5
+	github.com/mitchellh/go-testing-interface v1.0.0
+	github.com/nicksnyder/go-i18n v0.0.0-20190519163303-3c6ee9071cf0
+	github.com/oklog/run v1.0.0
+	github.com/pbnjay/pixfont v0.0.0-20190130005054-401bb7c6aee2
+	github.com/pborman/uuid v1.2.0
+	github.com/pelletier/go-toml v1.4.0
+	github.com/pkg/errors v0.8.1
+	github.com/pmezard/go-difflib v1.0.0
+	github.com/stretchr/testify v1.3.0
+	go.uber.org/atomic v1.4.0
+	go.uber.org/multierr v1.1.0
+	go.uber.org/zap v1.10.0
+	golang.org/x/crypto v0.0.0-20190513172903-22d7a77e9e5f
+	golang.org/x/net v0.0.0-20190514140710-3ec191127204
+	golang.org/x/sys v0.0.0-20190516110030-61b9204099cb
+	golang.org/x/text v0.3.2
+	google.golang.org/genproto v0.0.0-20190516172635-bb713bdc0e52
+	google.golang.org/grpc v1.20.1
+	gopkg.in/asn1-ber.v1 v1.0.0-20181015200546-f715ec2f112d
+	gopkg.in/natefinch/lumberjack.v2 v2.0.0
+	gopkg.in/yaml.v2 v2.2.2
+)
