+++ plugin.json
@@ -2,9 +2,15 @@
     "id": "emojigen",
     "name": "EmojiGen",
     "description": "Generate emoji.",
-    "version": "0.0.1",
-    "backend": {
-        "executable": "plugin"
+    "version": "0.1.0",
+    "min_server_version": "5.10.0",
+    "server": {
+        "executables": {
+            "linux-amd64": "server/dist/plugin-linux-amd64",
+            "darwin-amd64": "server/dist/plugin-darwin-amd64",
+            "windows-amd64": "server/dist/plugin-windows-amd64.exe"
+        },
+        "executable": ""
     },
     "settings_schema": {
         "settings": [
