+++ server/font.go
@@ -1,4 +1,4 @@
-package util
+package main
 
 import (
 	"bytes"
