+++ server/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/mattermost/mattermost-server/plugin"
+
+func main() {
+	plugin.ClientMain(&EmojigenPlugin{})
+}
