+++ server/manifest.go
@@ -1,0 +1,9 @@
+package main
+
+var manifest = struct {
+	Id      string
+	Version string
+}{
+	Id:      "emojigen",
+	Version: "0.1.0",
+}
